Move FileReader to webidl2js
