+++ lib/jsdom/browser/Window.js
@@ -12,8 +12,8 @@ const DOMException = require("../web-idl/DOMException");
 const btoa = require("abab").btoa;
 const atob = require("abab").atob;
 const idlUtils = require("../living/generated/utils");
-const createFileReader = require("../living/file-reader");
 const createXMLHttpRequest = require("../living/xmlhttprequest");
+const createFileReader = require("../living/generated/FileReader").createInterface;
 const Document = require("../living/generated/Document");
 const Navigator = require("../living/generated/Navigator");
 const reportException = require("../living/helpers/runtime-script-errors");
@@ -261,7 +261,10 @@ function Window(options) {
     return result;
   };
 
-  this.FileReader = createFileReader(this);
+  this.FileReader = createFileReader({
+    window: this
+  }).interface;
+
   this.XMLHttpRequest = createXMLHttpRequest(this);
 
   // TODO: necessary for Blob and FileReader due to different-globals weirdness; investigate how to avoid this.
