+++ scripts/webidl/convert.js
@@ -22,6 +22,7 @@ addDir("../../lib/jsdom/living/attributes");
 addDir("../../lib/jsdom/living/window");
 addDir("../../lib/jsdom/living/nodes");
 addDir("../../lib/jsdom/living/navigator");
+addDir("../../lib/jsdom/living/file-api");
 addDir("../../lib/jsdom/living/xhr");
 addDir("../../lib/jsdom/living/domparsing");
